pyproject.toml
[project]
name = "ninovanotifier"
version = "0.2.2"
description = "Telegram bot that notifies about new grades on ninova"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "plyer>=2.1.0",
    "pytelegrambotapi>=4.29.1",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "rich>=14.2.0",
    "cryptography>=41.0.0",
    "matplotlib>=3.10.8",
    "scipy>=1.17.0",
    "numpy>=2.4.1",
    "sqlalchemy>=2.0.48",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "pre-commit>=4.0.0",
    "ruff>=0.8.0",
    "detect-secrets>=1.5.0",
]

[tool.ruff]
target-version = "py312"
line-length = 100
exclude = [".venv", "build", "dist", "__pycache__"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort (import sorting)
    "B",      # flake8-bugbear (common pitfalls)
    "C4",     # flake8-comprehensions (better comprehensions)
    "UP",     # pyupgrade (modern Python syntax)
    "RET",    # flake8-return (return statement best practices)
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify (code simplification)
    "N",      # pep8-naming (naming conventions)
    "A",      # flake8-builtins (shadowing builtins)
    "LOG",    # flake8-logging (logging best practices)
    "G",      # flake8-logging-format
    "PT",     # flake8-pytest-style (test best practices)
    "PIE",    # flake8-pie (misc. lints)
    "ICN",    # flake8-import-conventions
    "RUF",    # Ruff-specific rules
    "PERF",   # Perflint (performance anti-patterns)
    "PTH",    # flake8-use-pathlib (pathlib over os.path)
    "RSE",    # flake8-raise (raise best practices)
    "TCH",    # flake8-type-checking (TYPE_CHECKING imports)
    "FURB",   # refurb (modern Python idioms)
]
ignore = [
    "E501",     # line too long (handled by formatter)
    "S",        # Security (Bandit)
    "T20",      # No print statements
    "TRY",      # Tryceratops
    "EM",       # Error message formatting
    "S101",     # Allow assert
    "RUF001",   # Ambiguous unicode character (allow ı, ö, ü, vb.)
    "RUF002",   # Ambiguous unicode character in docstring
    "RUF003",   # Ambiguous unicode character in comment
    "G004",     # logging-f-string
    "FBT001",   # Boolean-typed positional argument in function definition
    "FBT002",   # Boolean default positional argument in function definition
    "FBT003",   # Boolean positional value in function call
]


[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
addopts = "-v --tb=short"
